backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
PyMuPDF>=1.24.0
slowapi>=0.1.9
neo4j>=5.25.0
tavily-python>=0.5.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest-asyncio>=0.23
pytest-cov>=5.0
pytest-mock>=3.14
freezegun>=1.5
